pyproject.toml
[tool.poetry]
name = "lightbikeshop"
version = "0.0.0"
description = ""
authors = ["pharci <lesha.belikov2005@yandex.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django-admin-interface = "^0.29.4"
django = "^5.1.6"
django-rangefilter = "^0.9.1"
django-jsonform = "^2.23.2"
django-admin-logs = "^1.3.0"
django-admin-column-toggle = "^0.4.2"
django-nested-admin = "^4.1.1"
django-colorfield = "^0.12.0"
requests = "^2.32.3"
django-admin-sortable2 = "^2.2.8"
pyjwt = "^2.10.1"
cryptography = "^45.0.6"
social-auth-app-django = "^5.5.1"
openpyxl = "^3.1.5"
python-slugify = "^8.0.4"
dotenv = "^0.9.9"
django-debug-toolbar = "^6.0.0"
unidecode = "^1.4.0"
celery = "^5.5.3"
redis = "^6.4.0"
psycopg2 = "^2.9.10"
gunicorn = "^23.0.0"
python-json-logger = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
